rename table.ts to recordbatch.ts in preparation for merging latest
